requirements.txt
django
psycopg2-binary
cohere
python-dotenv
pgvector
pypdf
django-extensions
Werkzeug
django-allauth[socialaccount]
openai
django-debug-toolbar
mypy
django-stubs
anthropic
tenacity
uvicorn
whitenoise
typeguard
pydantic
channels
daphne
google-generativeai
django-storages[google]
django-storages[s3]
pytest
pytest-cov
pytest-django
boto3
websockets
debugpy
channels-redis
#djangorestframework

celery[redis]
#chunking experiments
#google-genai
#groq
#instructor

Pillow

google-generativeai
dotenv